[DBC] add extended proc flag support

+++ engine/item/special_effect.hpp
@@ -44,7 +44,7 @@ struct special_effect_t
   special_effect_e type;
   special_effect_source_e source;
   std::string name_str, trigger_str, encoding_str;
-  unsigned proc_flags_; /* Proc-by */
+  uint64_t proc_flags_; /* Proc-by */
   unsigned proc_flags2_; /* Proc-on (hit/damage/...) */
   stat_e stat;
   school_e school;
@@ -153,7 +153,7 @@ public:
   spell_t* initialize_resource_action() const;
 
   /* Accessors for driver specific features of the proc; some are also used for on-use effects */
-  unsigned proc_flags() const;
+  uint64_t proc_flags() const;
   unsigned proc_flags2() const;
   double ppm() const;
   double rppm() const;
